现在开始提取状态和代码模块化 一天做一个模块,多好
