+++ src/App.js
@@ -2,65 +2,16 @@ import React from 'react';
 import 'antd/dist/reset.css';
 import "./App.css"
 import { Layout, Space,Col, Divider, Row,Typography ,Button,Card} from 'antd';
-import { AudioOutlined } from '@ant-design/icons';
 import { Input } from 'antd';
 import logoPic from './img/logo192.png';
-import { MailOutlined, SettingOutlined,BookOutlined,LeftOutlined,RightOutlined   } from '@ant-design/icons';
-import { Menu } from 'antd';
+import {LeftOutlined,RightOutlined   } from '@ant-design/icons';
 import Exam from './Exam.js'
 import Foot from './Foot.js'
+import MainContent from './MainContent';
 const { Search } = Input;
 const { Header, Footer, Sider, Content } = Layout;
 const { Title, Paragraph, Text, Link } = Typography;
-const headerStyle = {
-  textAlign: 'center',
-  color: '#fff',
-  height: 64,
-  paddingInline: 50,
-  lineHeight: '64px',
-  backgroundColor: '#fafafa',
-};
-const contentStyle = {
-  textAlign: 'center',
-  minHeight: 120,
-  lineHeight: '120px',
-  color: '#fff',
-  backgroundColor: '#fafafa',
-};
-const qusetionStyle ={
-  textAlign: 'center',
-  minHeight: 120,
-  lineHeight: '120px',
-}
-const leftNavStyle = {
-  textAlign: 'left',
-  lineHeight: '80px',
-  color: '#fff',
-  backgroundColor: '#fafafa',
-}
-const siderStyle = {
-  textAlign: 'left',
-  lineHeight: '120px',
-  color: '#fff',
-  backgroundColor: '#fafafa',
-};
-const footerStyle = {
-  textAlign: 'center',
-  color: '#fff',
-  backgroundColor: '#fafafa',
-};
-const questionFooterStyle = {
-  textAlign:"right",
-  backgroundColor: '#fafafa',
-};
-const suffix = (
-  <AudioOutlined
-    style={{
-      fontSize: 16,
-      color: '#1890ff',
-    }}
-  />
-);
+
 
 const onSearch = (value) => console.log(value);
 
@@ -87,86 +38,7 @@ class Top extends React.Component{
   }
 }
 
-//导航栏
-function getItem(label, key, icon, children, type) {
-  return {
-    key,
-    icon,
-    children,
-    label,
-    type,
-  };
-}
-const items = [
-  getItem('Navigation One', 'sub1', <BookOutlined />, [
-    getItem('Item 1', 'g1', null, [getItem('Option 1', '1'), getItem('Option 2', '2')], 'group'),
-    getItem('Item 2', 'g2', null, [getItem('Option 3', '3'), getItem('Option 4', '4')], 'group'),
-  ]),
-  getItem('Navigation Two', 'sub2', <BookOutlined />, [
-    getItem('Option 5', '5'),
-    getItem('Option 6', '6'),
-    getItem('Submenu', 'sub3', null, [getItem('Option 7', '7'), getItem('Option 8', '8')]),
-  ]),
-  {
-    type: 'divider',
-  },
-  getItem('Navigation Three', 'sub4', <BookOutlined />, [
-    getItem('Option 9', '9'),
-    getItem('Option 10', '10'),
-    getItem('Option 11', '11'),
-    getItem('Option 12', '12'),
-  ]),
-  getItem('Group', 'grp', null, [getItem('Option 13', '13'), getItem('Option 14', '14')], 'group'),
-];
 
-class MainContent extends React.Component{
-  constructor(props){
-    super(props);
-  }
-  onNavClick = (e) => {
-    console.log('click ', e);
-  };
-  render(){
-    return(
-      <div className='mainContent'>
-        <Layout>
-          <Sider className='leftNav' style={leftNavStyle}>
-            <div className='leftNav'>
-              <Menu
-                  onClick={this.onNavClick.bind(this)}
-                  style={{
-                    width: 256,
-                  }}
-                  defaultSelectedKeys={['1']}
-                  defaultOpenKeys={['sub1']}
-                  mode="inline"
-                  items={items}
-              />
-            </div>
-          </Sider>
-          <Content style={contentStyle}>
-            <div className='map'>
-              
-            </div>
-          </Content>
-          <Sider  style={siderStyle}>
-            <div className='info'>
-              <Typography>
-                <Title level={2}>Introduction</Title>
-                <Paragraph >
-                  In the process of internal desktop applications development, many different design specs and
-                  implementations would be involved, which might cause designers and developers difficulties and
-                  duplication and reduce the efficiency of development.
-                </Paragraph>
-            </Typography>
-            </div>
-          </Sider>
-          </Layout>
-       
-      </div>
-    );
-  }
-}
 
 class Question extends React.Component{
   constructor(props){
@@ -174,10 +46,9 @@ class Question extends React.Component{
   }
   render(){
     return(
-      <div className='question'>
-        <Divider></Divider>
+      <Card id='questionCard' size="default" title="Question" bordered >
+        <div className='question'>
           <Typography>
-            <Title level={2} className='titile'>Introduction</Title>
             <Paragraph className='paragraph'>
               In the process of internal desktop applications development, many different design specs and
               implementations would be involved, which might cause designers and developers difficulties and
@@ -190,8 +61,9 @@ class Question extends React.Component{
             <Button icon={<RightOutlined />} size={'small'} />
           </Space>
           </div>
-          <Divider></Divider>
-      </div>
+          
+        </div>
+      </Card>
     );
   }
 }
@@ -204,6 +76,7 @@ class SecondContent extends React.Component{
     return(
       <div className='secondContent'>
         <Question/>
+        <Divider/>
         <Exam/>
       </div>
     );
@@ -216,10 +89,11 @@ const Page = () => (
       <Top/>
       <Content>
         <MainContent></MainContent>
+        <Divider/>
         <SecondContent></SecondContent>
       </Content>
+      <Divider/>
       <Foot></Foot>
-      
     </Layout>
   </div>
